requirements.txt
aniso8601==9.0.1
astroid==2.5.6
attrs==20.3.0
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
flask-mongoengine==1.0.0
flask-restplus==0.13.0
Flask-WTF==0.14.3
idna==3.1
importlib-metadata==4.0.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.23.0
pycodestyle==2.7.0
pylint==2.8.2
pymongo==3.11.3
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
Pillow==8.2.0